% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox_summary_new_UI.R
\name{toxSummary_app}
\alias{toxSummary_app}
\title{Run toxSummary app}
\usage{
toxSummary_app(
  database_path,
  studyid_file,
  save_file_path = NULL,
  where_to_run = "local"
)
}
\arguments{
\item{database_path}{Mandatory, character\cr
file path for database}

\item{studyid_file}{Mandatory, character\cr
file path for studyid}

\item{save_file_path}{Optional, character\cr
path where file will be saved when app create any files.
If NULL, file will be saved in  working directory}

\item{where_to_run}{Optional, character\cr
where app will be running. Default is "local". Should use "rsconnect"
when deploy to rsconnect.}
}
\value{
function run the app.
}
\description{
Run toxSummary app
}
\examples{
\dontrun{
toxSummary_app(database_path = "path/of/your/database.db",
studyid_file = "path/for/IND_studyid_mapping.csv", 
save_file_path = NULL,
where_to_run = "local")
}
}
