% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dose_pp.R
\name{get_pk_param}
\alias{get_pk_param}
\title{Title}
\usage{
get_pk_param(conn, studyid, pk_param = "AUCLST", sex_include = "ALL")
}
\arguments{
\item{studyid}{}

\item{pk_param}{}

\item{db}{}
}
\value{

}
\description{
Title
}
