% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxSummary.R
\name{toxsummary_app}
\alias{toxsummary_app}
\title{run toxSummary app}
\usage{
toxsummary_app(
  database_path,
  studyid_file,
  save_file_path = NULL,
  where_to_run = "local"
)
}
\arguments{
\item{database_path}{Mandatory/cr
file path for database}

\item{studyid_file}{Mandatory/cr
file path for studyid}

\item{save_file_path}{optional
path where created application file will be saved.
If NULL, file will be saved in  working directory}

\item{where_to_run}{Mandatory/cr
where run will be running. Default is local}
}
\value{
function run the app.
}
\description{
run toxSummary app
}
